azure_function/requirements.txt
azure-functions
azure-cosmos
requests